Refactor JWT package to modularize key functions
This refactor encapsulates key management for JWT signing and verification by:
- Introducing `ParseKeyFunc` and `SecuringKeyFn` to streamline key operations.
- Moving key-related functions to `jwt.go` to improve package modularity.
- Simplifying `Encrypt` and `Decrypt` function documentation and usage.
- Updating tests to align with the new function signatures and logic flow.
